requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiofiles
pybind11
uvicorn
requests
pillow
webpocket
argon2-cffi